# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API E-commerce Fashion",
        "description": "Hợp đồng API giữa Vue.js Frontend và Laravel Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/products/variants/search-sku": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Tìm kiếm biến thể sản phẩm theo mã SKU",
                "operationId": "88229386d5290417139e2eb601284bef",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "query",
                        "description": "Mã SKU cần tìm",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tìm kiếm thành công"
                    },
                    "400": {
                        "description": "Lỗi khi thiếu mã SKU"
                    }
                }
            }
        },
...
    },
...
}